requirements.txt
asgiref==3.5.2
click==8.1.3
Django==3.2
djangorestframework==3.14.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.23.5
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
